+++ src/Type/BotCommand.php
@@ -14,13 +14,18 @@ final readonly class BotCommand
     public function __construct(
         public string $command,
         public string $description,
+        public ?bool $isEphemeral = null,
     ) {}
 
     public function toRequestArray(): array
     {
-        return [
-            'command' => $this->command,
-            'description' => $this->description,
-        ];
+        return array_filter(
+            [
+                'command' => $this->command,
+                'description' => $this->description,
+                'is_ephemeral' => $this->isEphemeral,
+            ],
+            static fn(mixed $value): bool => $value !== null,
+        );
     }
 }