+++ tests/Type/BotCommandTest.php
@@ -8,7 +8,9 @@ use PHPUnit\Framework\TestCase;
 use Phptg\BotApi\ParseResult\ObjectFactory;
 use Phptg\BotApi\Type\BotCommand;
 
+use function PHPUnit\Framework\assertNull;
 use function PHPUnit\Framework\assertSame;
+use function PHPUnit\Framework\assertTrue;
 
 final class BotCommandTest extends TestCase
 {
@@ -18,11 +20,27 @@ final class BotCommandTest extends TestCase
 
         assertSame('start', $botCommand->command);
         assertSame('Start command', $botCommand->description);
+        assertNull($botCommand->isEphemeral);
 
         assertSame(
             [
                 'command' => 'start',
                 'description' => 'Start command',
+            ],
+            $botCommand->toRequestArray(),
+        );
+    }
+
+    public function testFull(): void
+    {
+        $botCommand = new BotCommand('start', 'Start command', true);
+
+        assertTrue($botCommand->isEphemeral);
+        assertSame(
+            [
+                'command' => 'start',
+                'description' => 'Start command',
+                'is_ephemeral' => true,
             ],
             $botCommand->toRequestArray(),
         );
@@ -37,5 +55,17 @@ final class BotCommandTest extends TestCase
 
         assertSame('start', $botCommand->command);
         assertSame('Start command', $botCommand->description);
+        assertNull($botCommand->isEphemeral);
+    }
+
+    public function testFromTelegramResultFull(): void
+    {
+        $botCommand = (new ObjectFactory())->create([
+            'command' => 'start',
+            'description' => 'Start command',
+            'is_ephemeral' => true,
+        ], null, BotCommand::class);
+
+        assertTrue($botCommand->isEphemeral);
     }
 }